created separated models

+++ iBeaconOccupancy/resources/scene_analysis.py
@@ -1,7 +1,8 @@
 from flask import Flask, request, jsonify
 from flask.ext.restful import Resource
 from flask_sqlalchemy import SQLAlchemy
-from iBeaconOccupancy.model import Beacons,db,Locations,BeaconLocations
+from iBeaconOccupancy.model.beacons  import *
+from ..machine_learning import machine_learning
 
 class deviceFullLogic(Resource):
     def get(self,device):
@@ -17,9 +18,13 @@ class deviceFullLogic(Resource):
        #deleting all rows in table beacons
       beacons =Beacons.query.filter_by(id_device=device).all()
       for loc in beacons:
+
         db.session.delete(loc)
      #deleting row in table location
+
       location = Locations.query.filter_by(id_device=device).first()
+      if location is None:
+          return "no location with this characteristic"
       db.session.delete(location)
       db.session.commit()
 
@@ -46,7 +51,7 @@ class deviceFullLogic(Resource):
 
 
       mytest=[{'e2c56db5-dffb-48d2-b060-d0f5a71096e0035' : 9.23,'e2c56db5-dffb-48d2-b060-d0f5a71096e000':1.93}]
-      prediction = learning_machine.find_best_room(test_data)
+      prediction = machine_learning.find_best_room(test_data)
 
       #adding beacon to the beacons table
       #db.session.merge(Beacons(device, req["id_beacon"],req["status"],req["power"],req["user"],datetime.now()))
